Refactor AcademiesDbContext to filter out closed trusts
