+++ CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/). To see an example from a mature product in the program [see the Complete products changelog that follows the same methodology](https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/blob/main/CHANGELOG.md).
 
-
 ## [Unreleased][unreleased]
+
+### Changed
+
+- Filter the Gias Groups (Trusts) at the context level so they will never show up in searches or when pages load
 
 - Updated Feedback link to reflect new feedback form for the product
 
@@ -15,7 +18,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Migrated the application to LTS .NET 8. (FIAT was on .NET 7 which Microsoft ended support for in May).
 - Updated text on "404 - Not Found" page to adhere to DfE design pattern and added the not found url to the support email template.
-
 
 [unreleased]:
   https://github.com/DFE-Digital/dfe-complete-conversions-transfers-and-changes/compare/production-2024-07-18.2517...HEAD
