+++ DfE.FindInformationAcademiesTrusts.Data.AcademiesDb/TrustProvider.cs
@@ -1,4 +1,5 @@
 using System.Diagnostics.CodeAnalysis;
+using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Contexts;
 using DfE.FindInformationAcademiesTrusts.Data.AcademiesDb.Factories;
 using Microsoft.EntityFrameworkCore;
